+++ packages/opa/src/funcs/executeDefaultPolicyWithInput.ts
@@ -3,13 +3,10 @@
  */
 
 import { OpaApiClientCore } from "../core.js";
-import {
-  encodeFormQuery as encodeFormQuery$,
-  encodeJSON as encodeJSON$,
-  encodeSimple as encodeSimple$,
-} from "../lib/encodings.js";
-import * as m$ from "../lib/matchers.js";
-import * as schemas$ from "../lib/schemas.js";
+import { encodeFormQuery, encodeJSON, encodeSimple } from "../lib/encodings.js";
+import * as M from "../lib/matchers.js";
+import { compactMap } from "../lib/primitives.js";
+import { safeParse } from "../lib/schemas.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { extractSecurity, resolveGlobalSecurity } from "../lib/security.js";
 import { pathToFunc } from "../lib/url.js";
@@ -25,18 +22,19 @@ import * as errors from "../sdk/models/errors/index.js";
 import { SDKError } from "../sdk/models/errors/sdkerror.js";
 import { SDKValidationError } from "../sdk/models/errors/sdkvalidationerror.js";
 import * as operations from "../sdk/models/operations/index.js";
+import { APICall, APIPromise } from "../types/async.js";
 import { Result } from "../types/fp.js";
 
 /**
  * Execute the default decision  given an input
  */
-export async function executeDefaultPolicyWithInput(
-  client$: OpaApiClientCore,
+export function executeDefaultPolicyWithInput(
+  client: OpaApiClientCore,
   input: components.Input,
   pretty?: boolean | undefined,
   acceptEncoding?: components.GzipAcceptEncoding | undefined,
   options?: RequestOptions,
-): Promise<
+): APIPromise<
   Result<
     operations.ExecuteDefaultPolicyWithInputResponse,
     | errors.ClientError
@@ -50,82 +48,123 @@ export async function executeDefaultPolicyWithInput(
     | ConnectionError
   >
 > {
+  return new APIPromise($do(
+    client,
+    input,
+    pretty,
+    acceptEncoding,
+    options,
+  ));
+}
+
+async function $do(
+  client: OpaApiClientCore,
+  input: components.Input,
+  pretty?: boolean | undefined,
+  acceptEncoding?: components.GzipAcceptEncoding | undefined,
+  options?: RequestOptions,
+): Promise<
+  [
+    Result<
+      operations.ExecuteDefaultPolicyWithInputResponse,
+      | errors.ClientError
+      | errors.ServerError
+      | SDKError
+      | SDKValidationError
+      | UnexpectedClientError
+      | InvalidRequestError
+      | RequestAbortedError
+      | RequestTimeoutError
+      | ConnectionError
+    >,
+    APICall,
+  ]
+> {
   const input$: operations.ExecuteDefaultPolicyWithInputRequest = {
+    input: input,
     pretty: pretty,
     acceptEncoding: acceptEncoding,
-    input: input,
   };
 
-  const parsed$ = schemas$.safeParse(
+  const parsed = safeParse(
     input$,
-    (value$) =>
+    (value) =>
       operations.ExecuteDefaultPolicyWithInputRequest$outboundSchema.parse(
-        value$,
+        value,
       ),
     "Input validation failed",
   );
-  if (!parsed$.ok) {
-    return parsed$;
+  if (!parsed.ok) {
+    return [parsed, { status: "invalid" }];
   }
-  const payload$ = parsed$.value;
-  const body$ = encodeJSON$("body", payload$.input, { explode: true });
+  const payload = parsed.value;
+  const body = encodeJSON("body", payload.input, { explode: true });
 
-  const path$ = pathToFunc("/")();
+  const path = pathToFunc("/")();
 
-  const query$ = encodeFormQuery$({
-    "pretty": payload$.pretty,
+  const query = encodeFormQuery({
+    "pretty": payload.pretty,
   });
 
-  const headers$ = new Headers({
+  const headers = new Headers(compactMap({
     "Content-Type": "application/json",
     Accept: "application/json",
-    "Accept-Encoding": encodeSimple$(
+    "Accept-Encoding": encodeSimple(
       "Accept-Encoding",
-      payload$["Accept-Encoding"],
+      payload["Accept-Encoding"],
       { explode: false, charEncoding: "none" },
     ),
-  });
+  }));
 
-  const bearerAuth$ = await extractSecurity(client$.options$.bearerAuth);
-  const security$ = bearerAuth$ == null ? {} : { bearerAuth: bearerAuth$ };
+  const secConfig = await extractSecurity(client._options.bearerAuth);
+  const securityInput = secConfig == null ? {} : { bearerAuth: secConfig };
+  const requestSecurity = resolveGlobalSecurity(securityInput);
+
   const context = {
+    baseURL: options?.serverURL ?? client._baseURL ?? "",
     operationID: "executeDefaultPolicyWithInput",
     oAuth2Scopes: [],
-    securitySource: client$.options$.bearerAuth,
-  };
-  const securitySettings$ = resolveGlobalSecurity(security$);
 
-  const requestRes = client$.createRequest$(context, {
-    security: securitySettings$,
+    resolvedSecurity: requestSecurity,
+
+    securitySource: client._options.bearerAuth,
+    retryConfig: options?.retries
+      || client._options.retryConfig
+      || { strategy: "none" },
+    retryCodes: options?.retryCodes || ["429", "500", "502", "503", "504"],
+  };
+
+  const requestRes = client._createRequest(context, {
+    security: requestSecurity,
     method: "POST",
-    path: path$,
-    headers: headers$,
-    query: query$,
-    body: body$,
-    timeoutMs: options?.timeoutMs || client$.options$.timeoutMs || -1,
+    baseURL: options?.serverURL,
+    path: path,
+    headers: headers,
+    query: query,
+    body: body,
+    timeoutMs: options?.timeoutMs || client._options.timeoutMs || -1,
   }, options);
   if (!requestRes.ok) {
-    return requestRes;
+    return [requestRes, { status: "invalid" }];
   }
-  const request$ = requestRes.value;
+  const req = requestRes.value;
 
-  const doResult = await client$.do$(request$, {
+  const doResult = await client._do(req, {
     context,
     errorCodes: ["400", "404", "4XX", "500", "5XX"],
-    retryConfig: options?.retries
-      || client$.options$.retryConfig,
-    retryCodes: options?.retryCodes || ["429", "500", "502", "503", "504"],
+    retryConfig: context.retryConfig,
+    retryCodes: context.retryCodes,
   });
   if (!doResult.ok) {
-    return doResult;
+    return [doResult, { status: "request-error", request: req }];
   }
   const response = doResult.value;
 
-  const responseFields$ = {
-    HttpMeta: { Response: response, Request: request$ },
+  const responseFields = {
+    HttpMeta: { Response: response, Request: req },
   };
 
-  const [result$] = await m$.match<
+  const [result] = await M.match<
     operations.ExecuteDefaultPolicyWithInputResponse,
     | errors.ClientError
     | errors.ServerError
@@ -137,18 +176,19 @@ export async function executeDefaultPolicyWithInput(
     | RequestTimeoutError
     | ConnectionError
   >(
-    m$.json(
+    M.json(
       200,
       operations.ExecuteDefaultPolicyWithInputResponse$inboundSchema,
       { hdrs: true, key: "result" },
     ),
-    m$.jsonErr([400, 404], errors.ClientError$inboundSchema),
-    m$.jsonErr(500, errors.ServerError$inboundSchema),
-    m$.fail(["4XX", "5XX"]),
-  )(response, request$, { extraFields: responseFields$ });
-  if (!result$.ok) {
-    return result$;
+    M.jsonErr([400, 404], errors.ClientError$inboundSchema),
+    M.jsonErr(500, errors.ServerError$inboundSchema),
+    M.fail("4XX"),
+    M.fail("5XX"),
+  )(response, req, { extraFields: responseFields });
+  if (!result.ok) {
+    return [result, { status: "complete", request: req, response }];
   }
 
-  return result$;
+  return [result, { status: "complete", request: req, response }];
 }