+++ packages/opa/src/lib/security.ts
@@ -3,6 +3,13 @@
  */
 
 import * as components from "../sdk/models/components/index.js";
+type OAuth2PasswordFlow = {
+  username: string;
+  password?: string | undefined;
+  clientID: string;
+  clientSecret?: string | undefined;
+  tokenURL: string;
+};
 
 export enum SecurityErrorCode {
   Incomplete = "incomplete",
@@ -37,6 +44,7 @@ export type SecurityState = {
   headers: Record<string, string>;
   queryParams: Record<string, string>;
   cookies: Record<string, string>;
+  oauth2: ({ type: "password" } & OAuth2PasswordFlow) | { type: "none" };
 };
 
 type SecurityInputBasic = {
@@ -73,7 +81,18 @@ type SecurityInputOAuth2 = {
 
 type SecurityInputOAuth2ClientCredentials = {
   type: "oauth2:client_credentials";
-  value: string | null | undefined;
+  value:
+    | { clientID?: string | undefined; clientSecret?: string | undefined }
+    | null
+    | undefined;
+};
+
+type SecurityInputOAuth2PasswordCredentials = {
+  type: "oauth2:password";
+  value:
+    | string
+    | null
+    | undefined;
   fieldName: string;
 };
 
@@ -89,6 +108,7 @@ export type SecurityInput =
   | SecurityInputAPIKey
   | SecurityInputOAuth2
   | SecurityInputOAuth2ClientCredentials
+  | SecurityInputOAuth2PasswordCredentials
   | SecurityInputOIDC
   | SecurityInputCustom;
 
@@ -96,10 +116,11 @@ export function resolveSecurity(
   ...options: SecurityInput[][]
 ): SecurityState | null {
   const state: SecurityState = {
-    basic: { username: "", password: "" },
+    basic: {},
     headers: {},
     queryParams: {},
     cookies: {},
+    oauth2: { type: "none" },
   };
 
   const option = options.find((opts) => {
@@ -110,6 +131,12 @@ export function resolveSecurity(
         return o.value.username != null || o.value.password != null;
       } else if (o.type === "http:custom") {
         return null;
+      } else if (o.type === "oauth2:password") {
+        return (
+          typeof o.value === "string" && !!o.value
+        );
+      } else if (o.type === "oauth2:client_credentials") {
+        return o.value.clientID != null || o.value.clientSecret != null;
       } else if (typeof o.value === "string") {
         return !!o.value;
       } else {
@@ -152,6 +179,9 @@ export function resolveSecurity(
       case "oauth2":
         applyBearer(state, spec);
         break;
+      case "oauth2:password":
+        applyBearer(state, spec);
+        break;
       case "oauth2:client_credentials":
         break;
       case "openIdConnect":
@@ -179,9 +209,13 @@ function applyBasic(
 
 function applyBearer(
   state: SecurityState,
-  spec: SecurityInputBearer | SecurityInputOAuth2 | SecurityInputOIDC,
+  spec:
+    | SecurityInputBearer
+    | SecurityInputOAuth2
+    | SecurityInputOIDC
+    | SecurityInputOAuth2PasswordCredentials,
 ) {
-  if (spec.value == null) {
+  if (typeof spec.value !== "string" || !spec.value) {
     return;
   }
 
@@ -192,6 +226,7 @@ function applyBearer(
 
   state.headers[spec.fieldName] = value;
 }
+
 export function resolveGlobalSecurity(
   security: Partial<components.Security> | null | undefined,
 ): SecurityState | null {