+++ packages/opa/src/sdk/models/errors/clienterror.ts
@@ -3,6 +3,9 @@
  */
 
 import * as z from "zod";
+import { safeParse } from "../../../lib/schemas.js";
+import { Result as SafeParseResult } from "../../../types/fp.js";
+import { SDKValidationError } from "./sdkvalidationerror.js";
 
 export type Location = {
   file: string;
@@ -91,6 +94,20 @@ export namespace Location$ {
   export type Outbound = Location$Outbound;
 }
 
+export function locationToJSON(location: Location): string {
+  return JSON.stringify(Location$outboundSchema.parse(location));
+}
+
+export function locationFromJSON(
+  jsonString: string,
+): SafeParseResult<Location, SDKValidationError> {
+  return safeParse(
+    jsonString,
+    (x) => Location$inboundSchema.parse(JSON.parse(x)),
+    `Failed to parse 'Location' from JSON`,
+  );
+}
+
 /** @internal */
 export const Errors$inboundSchema: z.ZodType<Errors, z.ZodTypeDef, unknown> = z
   .object({
@@ -128,6 +145,20 @@ export namespace Errors$ {
   export const outboundSchema = Errors$outboundSchema;
   /** @deprecated use `Errors$Outbound` instead. */
   export type Outbound = Errors$Outbound;
+}
+
+export function errorsToJSON(errors: Errors): string {
+  return JSON.stringify(Errors$outboundSchema.parse(errors));
+}
+
+export function errorsFromJSON(
+  jsonString: string,
+): SafeParseResult<Errors, SDKValidationError> {
+  return safeParse(
+    jsonString,
+    (x) => Errors$inboundSchema.parse(JSON.parse(x)),
+    `Failed to parse 'Errors' from JSON`,
+  );
 }
 
 /** @internal */